User data fetching moved to higher order component(page) for todo details component
